refactor: redis cache thread - redis cache 로직 수정

+++ bookmaru-application/src/main/kotlin/plain/bookmaru/common/command/PageCommand.kt
@@ -4,6 +4,6 @@ data class PageCommand(
     val page: Int,
     val size: Int
 ) {
-    val offset: Long
-        get() = (page * size).toLong()
+    val offset: Int
+        get() = (page * size)
 }
